# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManualPageConnection": {
        "properties": {
          "page_access_token": {
            "maxLength": 4096,
            "minLength": 20,
            "title": "Page Access Token",
            "type": "string"
          }
        },
        "required": [
          "page_access_token"
        ],
        "title": "ManualPageConnection",
        "type": "object"
      },
...
      "OAuthStart": {
        "properties": {
          "authorization_url": {
            "title": "Authorization Url",
            "type": "string"
          }
        },
        "required": [
          "authorization_url"
        ],
        "title": "OAuthStart",
        "type": "object"
      },
      "PageChoice": {
        "properties": {
          "can_moderate": {
            "title": "Can Moderate",
            "type": "boolean"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "page_name": {
            "title": "Page Name",
            "type": "string"
          },
          "tasks": {
            "items": {
              "type": "string"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "page_id",
          "page_name",
          "tasks",
          "can_moderate"
        ],
        "title": "PageChoice",
        "type": "object"
      },
      "PageChoices": {
        "properties": {
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageChoice"
            },
            "title": "Pages",
            "type": "array"
          }
        },
        "required": [
          "pages"
        ],
        "title": "PageChoices",
        "type": "object"
      },
      "PageConnection": {
        "properties": {
          "can_moderate": {
            "title": "Can Moderate",
            "type": "boolean"
          },
          "connected_at": {
            "format": "date-time",
            "title": "Connected At",
            "type": "string"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "method": {
            "enum": [
              "FACEBOOK_LOGIN",
              "MANUAL_TOKEN"
            ],
            "title": "Method",
            "type": "string"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "page_name": {
            "title": "Page Name",
            "type": "string"
          },
          "state": {
            "const": "CONNECTED",
            "default": "CONNECTED",
            "title": "State",
            "type": "string"
          },
          "tasks": {
            "items": {
              "type": "string"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "page_id",
          "page_name",
          "method",
          "tasks",
          "can_moderate",
          "connected_at",
          "last_synced_at"
        ],
        "title": "PageConnection",
        "type": "object"
      },
      "PageConnections": {
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/PageConnection"
            },
            "title": "Connections",
            "type": "array"
          },
          "page_limit": {
            "title": "Page Limit",
            "type": "integer"
          },
          "plan": {
            "enum": [
              "STARTER",
              "GROWTH"
            ],
            "title": "Plan",
            "type": "string"
          }
        },
        "required": [
          "plan",
          "page_limit",
          "connections"
        ],
        "title": "PageConnections",
        "type": "object"
      },
...
      "SelectPage": {
        "properties": {
          "page_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Page Id",
            "type": "string"
          }
        },
        "required": [
          "page_id"
        ],
        "title": "SelectPage",
        "type": "object"
      },
...
      "SyncResult": {
        "properties": {
          "fetched": {
            "title": "Fetched",
            "type": "integer"
          },
          "imported": {
            "title": "Imported",
            "type": "integer"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "page_name": {
            "title": "Page Name",
            "type": "string"
          }
        },
        "required": [
          "fetched",
          "imported",
          "page_id",
          "page_name",
          "last_synced_at"
        ],
        "title": "SyncResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "KCMS Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/facebook/connections": {
      "get": {
        "operationId": "listFacebookConnections",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageConnections"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Connections"
      }
    },
    "/api/v1/facebook/connections/manual": {
      "post": {
        "operationId": "connectFacebookPageManually",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPageConnection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageConnection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect Manually"
      }
    },
    "/api/v1/facebook/connections/{page_id}": {
      "delete": {
        "operationId": "disconnectFacebookPage",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disconnect"
      }
    },
    "/api/v1/facebook/connections/{page_id}/sync": {
      "post": {
        "description": "Pull comments from the connected Page into this workspace.\n\nRe-syncing is safe: the provider's comment id is the primary key, so an\nalready-imported comment keeps its verdict, actions and corrections.",
        "operationId": "syncFacebookComments",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Comments"
      }
    },
    "/api/v1/facebook/oauth/sessions/{state}": {
      "get": {
        "operationId": "listFacebookPageChoices",
        "parameters": [
          {
            "in": "path",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageChoices"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Page Choices"
      }
    },
    "/api/v1/facebook/oauth/sessions/{state}/selection": {
      "post": {
        "operationId": "selectFacebookPage",
        "parameters": [
          {
            "in": "path",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectPage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageConnection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Page"
      }
    },
    "/api/v1/facebook/oauth/start": {
      "post": {
        "operationId": "startFacebookAuthorization",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthStart"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Authorization"
      }
    },
...
  }
}